Move Marketplace to single recorded test class
