+++ tools/Azure.Mcp.Tools.Marketplace/tests/Azure.Mcp.Tools.Marketplace.Tests/MarketplaceCommandTests.cs
@@ -9,11 +9,80 @@ using Xunit;
 
 namespace Azure.Mcp.Tools.Marketplace.Tests;
 
-public sealed class ProductListCommandTests(ITestOutputHelper output, TestProxyFixture fixture, LiveServerFixture liveServerFixture)
+public sealed class MarketplaceCommandTests(ITestOutputHelper output, TestProxyFixture fixture, LiveServerFixture liveServerFixture)
     : RecordedCommandTestsBase(output, fixture, liveServerFixture)
 {
+    private const string ProductKey = "product";
     private const string ProductsKey = "products";
+    private const string ProductId = "test_test_pmc2pc1.vmsr_uat_beta";
     private const string Language = "en";
+
+    #region product_get
+
+    [Fact]
+    public async Task Should_get_marketplace_product()
+    {
+        var result = await CallToolAsync(
+            "marketplace_product_get",
+            new()
+            {
+                { "subscription", Settings.SubscriptionId },
+                { "product-id", ProductId }
+            });
+
+        var product = result.AssertProperty(ProductKey);
+        Assert.Equal(JsonValueKind.Object, product.ValueKind);
+
+        var id = product.AssertProperty("uniqueProductId");
+        Assert.Equal(JsonValueKind.String, id.ValueKind);
+        Assert.Contains(ProductId, id.GetString());
+    }
+
+    [Fact]
+    public async Task Should_get_marketplace_product_with_language_option()
+    {
+        var result = await CallToolAsync(
+            "marketplace_product_get",
+            new()
+            {
+                { "subscription", Settings.SubscriptionId },
+                { "product-id", ProductId },
+                { "language", Language }
+            });
+
+        var product = result.AssertProperty(ProductKey);
+        Assert.Equal(JsonValueKind.Object, product.ValueKind);
+
+        var id = product.AssertProperty("uniqueProductId");
+        Assert.Equal(JsonValueKind.String, id.ValueKind);
+        Assert.Contains(ProductId, id.GetString());
+    }
+
+    [Fact]
+    public async Task Should_get_marketplace_product_with_multiple_options()
+    {
+        var result = await CallToolAsync(
+            "marketplace_product_get",
+            new()
+            {
+                { "subscription", Settings.SubscriptionId },
+                { "product-id", ProductId },
+                { "language", Language },
+                { "include-hidden-plans", true },
+                { "include-service-instruction-templates", true }
+            });
+
+        var product = result.AssertProperty(ProductKey);
+        Assert.Equal(JsonValueKind.Object, product.ValueKind);
+
+        var id = product.AssertProperty("uniqueProductId");
+        Assert.Equal(JsonValueKind.String, id.ValueKind);
+        Assert.Contains(ProductId, id.GetString());
+    }
+
+    #endregion
+
+    #region product_list
 
     [Fact]
     public async Task Should_list_marketplace_products()
@@ -220,4 +289,6 @@ public sealed class ProductListCommandTests(ITestOutputHelper output, TestProxyF
         product.AssertProperty("uniqueProductId");
         product.AssertProperty("displayName");
     }
+
+    #endregion
 }
