+++ tools/Azure.Mcp.Tools.Marketplace/tests/Azure.Mcp.Tools.Marketplace.Tests/assets.json
@@ -2,5 +2,5 @@
   "AssetsRepo": "Azure/azure-sdk-assets",
   "AssetsRepoPrefixPath": "",
   "TagPrefix": "Azure.Mcp.Tools.Marketplace.Tests",
-  "Tag": "Azure.Mcp.Tools.Marketplace.Tests_3591345f11"
+  "Tag": "Azure.Mcp.Tools.Marketplace.Tests_ee5ae6dcb2"
 }
